backend/requirements.txt
flask
flask-WTF
Jinja2
sqlalchemy
wtforms
passlib
flask_restful
flask_security
flask_admin
MarkupSafe
itsdangerous
flask-Mail
flask_restful
flask_debugtoolbar
flask_jwt_extended
flask_socketio
eventlet
redis
flask_login
flask_mail
flask_sqlalchemy
client
suds-py3
requests
zeep
flask_admin
flask_security
pillow
psycopg2-binary
flask_session
flask_cors
gunicorn
flask-restplus
flask_migrate
flask_script
rejson
gevent-websocket
